+++ misc/gene_ontology_utils.py
@@ -1,0 +1,13 @@
+""" Utilities for working with the Gene Ontology (GO)
+
+Please see the GO website for more details (http://www.geneontology.org/).
+
+These utilities are often designed to work with `goatools`.
+
+* paper: https://www.nature.com/articles/s41598-018-28948-z
+* git repository: https://github.com/tanghaibao/goatools
+"""
+
+BIOLOGICAL_PROCESS_ROOT_TERM = 'GO:0008150'
+CELLULAR_COMPONENT_ROOT_TERM = 'GO:0005575'
+MOLECULAR_FUNCTION_ROOT_TERM = 'GO:0003674'